# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
}

export interface Patient {
  id: number;
  name: string;
  age: number;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  bloodType?: string | null;
  notes?: string | null;
  lastVisitDate?: string | null;
  createdAt: string;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  confirmed: "confirmed",
  pending: "pending",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: number;
  patientId: number;
  date: string;
  time: string;
  reason: string;
  status: AppointmentStatus;
  notes?: string | null;
  createdAt: string;
}

export type PatientWithHistory = Patient & {
  appointments: Appointment[];
};

export interface PatientListResponse {
  patients: Patient[];
  total: number;
}

export interface CreatePatientBody {
  name: string;
  age: number;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  bloodType?: string | null;
  notes?: string | null;
}

export interface UpdatePatientBody {
  name?: string;
  age?: number;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  bloodType?: string | null;
  notes?: string | null;
}

export type AppointmentWithPatient = Appointment & {
  patient: Patient;
};

export type CreateAppointmentBodyStatus =
  (typeof CreateAppointmentBodyStatus)[keyof typeof CreateAppointmentBodyStatus];

export const CreateAppointmentBodyStatus = {
  confirmed: "confirmed",
  pending: "pending",
  cancelled: "cancelled",
} as const;

export interface CreateAppointmentBody {
  patientId: number;
  date: string;
  time: string;
  reason: string;
  status?: CreateAppointmentBodyStatus;
  notes?: string | null;
}

export type UpdateAppointmentBodyStatus =
  (typeof UpdateAppointmentBodyStatus)[keyof typeof UpdateAppointmentBodyStatus];

export const UpdateAppointmentBodyStatus = {
  confirmed: "confirmed",
  pending: "pending",
  cancelled: "cancelled",
} as const;

export interface UpdateAppointmentBody {
  patientId?: number;
  date?: string;
  time?: string;
  reason?: string;
  status?: UpdateAppointmentBodyStatus;
  notes?: string | null;
}

export interface DashboardSummary {
  todayAppointmentsCount: number;
  patientsSeen: number;
  totalPatients: number;
  upcomingCount: number;
  confirmedCount: number;
  pendingCount: number;
  cancelledCount: number;
}

export type ListPatientsParams = {
  /**
   * Search by name or contact
   */
  search?: string;
  limit?: number;
  offset?: number;
};

export type ListAppointmentsParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
  /**
   * Filter by patient
   */
  patientId?: number;
  status?: ListAppointmentsStatus;
  startDate?: string;
  endDate?: string;
};

export type ListAppointmentsStatus =
  (typeof ListAppointmentsStatus)[keyof typeof ListAppointmentsStatus];

export const ListAppointmentsStatus = {
  confirmed: "confirmed",
  pending: "pending",
  cancelled: "cancelled",
} as const;

export type GetUpcomingAppointmentsParams = {
  limit?: number;
};

export type GetRecentPatientsParams = {
  limit?: number;
};